pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "life-os"
version = "0.1.0"
description = "A personal control plane that reasons across your life and safely executes decisions"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["config", "src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
